Updating REST endpoints
-"/users" moved to "/api/users"
-"/api/**" does not require authentication
-REST Update utilizes UpdateUserDto
-TO-DO allow update password & non-admin updates
-added "authenticate(User,String)" to UserService
-REST Delete requires user password
-list-users.html added "Id" column to table

+++ src/main/java/com/jeffreyghj/springusers/rest/UserRestController.java
@@ -2,6 +2,8 @@ package com.jeffreyghj.springusers.rest;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,16 +14,23 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.jeffreyghj.springusers.dao.RoleRepository;
+import com.jeffreyghj.springusers.dto.PasswordDto;
+import com.jeffreyghj.springusers.dto.UpdateUserDto;
+import com.jeffreyghj.springusers.dto.UserDto;
 import com.jeffreyghj.springusers.entity.User;
 import com.jeffreyghj.springusers.service.UserService;
 
 //@RestController
 @RestController
-@RequestMapping("/")
+@RequestMapping("/api")
 public class UserRestController {
 	
 	@Autowired
 	private UserService userService;
+	
+	@Autowired
+	private RoleRepository roleRepository;
 	
 	@GetMapping("/users")
 	public List<User> findAllUsers() {
@@ -40,32 +49,39 @@ public class UserRestController {
 	}
 	
 	@PostMapping("/users")
-	public User createUser(@RequestBody User theUser) {
+	public User createUser(@Valid @RequestBody UserDto theUserDto) {
 		
-		// set id to 0 to force save of new item even if another id is given
-		// 0 = "INSERT"
-		theUser.setId((long) 0); 
+//		// set id to 0 to force save of new item even if another id is given
+//		// 0 = "INSERT"
+//		theUser.setId((long) 0); 
+//		theUser.setRoles(Arrays.asList(roleRepository.findRoleByName("USER")));
+//		userService.save(theUser);
+//		return theUser;
 		
-		userService.save(theUser);
-		
-		return theUser;
+		return userService.createNewUser(theUserDto);
 	}
 	
 	// Update
 	@PutMapping("/users")
-	public User saveUser(@RequestBody User theUser) {
-		userService.save(theUser);
-		return theUser;
+	public User saveUser(@Valid @RequestBody UpdateUserDto theUser) {
+		//userService.save(theUser);
+		return userService.updateUser(theUser);
 	}
 	
 	// To Delete through HTTP DELETE 
 	@DeleteMapping("/users/{userId}")
-	public String deleteUser(@PathVariable Long userId) {
+	public String deleteUser(@PathVariable Long userId, @RequestBody PasswordDto passwordDto) {
 	
+		System.out.println("Password given: " + passwordDto.getPassword());
+		
 		User tempUser = userService.findById(userId);
 		
 		if ( tempUser == null ) {
 			throw new RuntimeException("User id not found: " + userId);
+		}
+		
+		if ( !userService.authenticate(tempUser, passwordDto.getPassword()) ) {
+			throw new RuntimeException("Incorrect password - delete not authorized");
 		}
 		
 		userService.deleteById(userId);
